Changed limit to only 3 student records.  Changed StudentInformation struct's latitude and longitude properties to be of type Double instead of Float so they agree with the data types now visible in the NSDictionary in the debugger.

+++ OnTheMap/StudentInformation.swift
@@ -16,8 +16,8 @@ struct StudentInformation{
     var lastName  : String = ""
     var mapString : String = ""
     var mediaURL  : String = ""
-    var latitude  : Float  = 0
-    var longitude : Float  = 0
+    var latitude  : Double  = 0
+    var longitude : Double  = 0
     var objectId  : String = ""
     
     //var updatedAt:  String
